+++ packages/platform/src/lib/content/marked/marked-content-highlighter.ts
@@ -3,5 +3,5 @@ export interface MarkedContentHighlighter {
 }
 
 export abstract class MarkedContentHighlighter {
-  abstract getHighlightExtension(): import('marked').marked.MarkedExtension;
+  abstract getHighlightExtension(): import('marked').MarkedExtension;
 }