+++ packages/vite-plugin-angular/src/lib/angular-storybook-plugin.ts
@@ -5,13 +5,13 @@ export function angularStorybookPlugin() {
       if (code.includes('"@storybook/angular"')) {
         return code.replace(
           /\"@storybook\/angular\"/g,
-          '"@storybook/angular/dist/client"',
+          '"@storybook/angular/dist/client/index.js"',
         );
       }
       if (code.includes("'@storybook/angular'")) {
         return code.replace(
           /\'@storybook\/angular\'/g,
-          "'@storybook/angular/dist/client'",
+          "'@storybook/angular/dist/client/index.js'",
         );
       }